video-processing-demo/requirements.txt
opencv-python
numpy
moviepy
torch
torchvision
mtcnn
openai-whisper
pyaudio
ffmpeg-python
tqdm 